requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Deep Learning
torch>=1.12.0
torchvision>=0.13.0

# Federated Learning
flwr>=1.0.0

# Privacy
opacus>=1.3.0
cryptography>=38.0.0

# NLP for product classification
transformers>=4.20.0
sentence-transformers>=2.2.0

# Causal Inference
statsmodels>=0.13.0
econml>=0.13.0
dowhy>=0.8

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
shap>=0.41.0

# Utilities
pyyaml>=6.0
tqdm>=4.64.0
joblib>=1.1.0
click>=8.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development
black>=22.0.0
flake8>=4.0.0
mypy>=0.950